Changes as requested by aeon for pull request to be merged.
