+++ config/params.ini
@@ -1,6 +1,6 @@
 ; There is detailed documentation for each parameter in the README.md
 [general]
-name=Botty Barbarian
+name=Botty
 monitor=0
 custom_message_hook=
 logg_lvl=debug
@@ -20,7 +20,7 @@ saved_games_folder=
 run_trav=0
 run_pindle=1
 run_eldritch=1
-run_shenk=1
+run_shenk=0
 run_nihlatak=0
 
 [char]
@@ -28,8 +28,8 @@ run_nihlatak=0
 ; ==== Mandatory Fields ====
 ; ==========================
 ; These configs have to be alligned with your d2r settings and char build
-type=barbarian
-casting_frames=9
+type=sorceress
+casting_frames=11
 num_loot_columns=5
 show_items=alt
 inventory_screen=i
@@ -37,7 +37,7 @@ force_move=e
 tp=f9
 belt_rows=4
 ;note: stand_still can not be the default "shift" as it would interfere with merc healing
-stand_still=ctrl
+stand_still=capslock
 ;note: this is different from the default hotkey as "~" is for many keyboards not reachable without also pressing altgr
 show_belt=k
 potion1=1
@@ -45,19 +45,20 @@ potion2=2
 potion3=3
 potion4=4
 cta_available=0
-weapon_switch=w
-battle_orders=f1
-battle_command=f2
+weapon_switch=x
+battle_orders=f7
+battle_command=f8
 ; ==========================
 ; ==== Optional configs ====
 ; ==========================
 stash_gold=1
-gold_trav_only=1
+gold_trav_only=0
 use_merc=1
-atk_len_trav=10.0
-atk_len_pindle=8.0
-atk_len_eldritch=8.0
-atk_len_shenk=10.0
+; Attack length for barbarians should be as high as 8-10 and even 10-12 for trav/shenk
+atk_len_trav=3.0
+atk_len_pindle=3.0
+atk_len_eldritch=3.0
+atk_len_shenk=4.0
 atk_len_nihlatak=4.0
 take_health_potion=0.8
 take_mana_potion=0.6
@@ -72,7 +73,7 @@ belt_hp_columns=1
 belt_mp_columns=1
 pre_buff_every_run=0
 
-; Supported runs: Trav, Pindle, Eld, Shenk, Nihlatak
+; Supported runs: Trav, Pindle, Eld, Shenk, Nihlathak
 [sorceress]
 teleport=f1
 skill_left=f2
@@ -83,7 +84,7 @@ thunder_storm=
 telekinesis=
 static=
 
-; Supported runs: Trav, Pindle, Eld, Shenk, Nihlatak (tele only)
+; Supported runs: Trav, Pindle, Eld, Shenk, Nihlathak (tele only)
 [hammerdin]
 teleport=f1
 concentration=f2
