+++ src/char/i_char.py
@@ -172,40 +172,6 @@ class IChar:
             else:
                 Logger.warning("Turns out weapon switch just took a long time. You ever considered getting a new internet provider or to upgrade your pc?")
 
-    def _do_hork(self, hork_time: int):
-        # Outburst's fine work below
-        # save current skill img
-        wait(0.1)
-        skill_before = cut_roi(self._screen.grab(), self._config.ui_roi["skill_right"])
-        wait(0.5)
-        keyboard.send(self._char_config["weapon_switch"])
-        wait(0.5)
-        #hork
-        if self._skill_hotkeys["find_item"]:
-            keyboard.send(self._skill_hotkeys["find_item"])
-            wait(0.5, 0.15)
-        mouse.move(637, 354)
-        wait(0.5, 0.15)
-        mouse.press(button="right")
-        wait(hork_time)
-        mouse.release(button="right")
-        wait(1)
-        keyboard.send(self._char_config["weapon_switch"])
-        wait(0.5, 0.15)
-        # Make sure that we are back at the previous skill
-        skill_after = cut_roi(self._screen.grab(), self._config.ui_roi["skill_right"])
-        _, max_val, _, _ = cv2.minMaxLoc(cv2.matchTemplate(skill_after, skill_before, cv2.TM_CCOEFF_NORMED))
-        if max_val < 0.96:
-            Logger.warning("Failed to switch weapon, try again")
-            wait(1.2)
-            skill_after = cut_roi(self._screen.grab(), self._config.ui_roi["skill_right"])
-            _, max_val, _, _ = cv2.minMaxLoc(cv2.matchTemplate(skill_after, skill_before, cv2.TM_CCOEFF_NORMED))
-            if max_val < 0.96:
-                keyboard.send(self._char_config["weapon_switch"])
-                wait(0.4)
-            else:
-                Logger.warning("Turns out weapon switch just took a long time. You ever considered getting a new internet provider or to upgrade your pc?")
-
     def pre_buff(self):
         pass
